mypyserver/requirements.txt
Pillow==8.4.0
opencv-python==4.5.5.62
numpy==1.21.5
tensorflow==2.7.0